20260821 - Move renaming to Configuration > This node
Renaming sat on the home page, which was the wrong home for it twice over. The
home page is a status view, and everything else that saves on its own rather
than with the config form already lives under Administration on the config
page, next to SSH access and Cloud services.

It is now the first entry in that group, deliberately outside #configForm so
there is no implication that Apply changes has anything to do with it. The
section also shows the node's own ret<node_id>.local, which is the address
worth bookmarking: owl.local is shared by every node and may reach any of them.

Rebuilt on the real field markup while moving it. The first version invented
cfg-field and hint classes that do not exist in the stylesheet; it now uses the
cfg-row / label-col / help / cfg-readonly pattern the rest of the page uses.
